Pattern History
Added pattern history so that the edge conditions where measurement or
generation jitter don't cause rapid flickering between output patterns.
Accomplished via a adding a ring buffer to store the last 5 measurements
and unless they all are the same display the last (oldest) pattern. Keep
pushing the latest (youngest) measument to the front of the ring.