Interrupt Fix - added pattern stability
Interrupt wasn't actually being disabled in previous code so with the
higher refresh rate of the motor output on the roboRIO it was causing
unpredictable behavior. Also changed the way the patternHistory data was
being tested and what to do when the pattern requested wasn't stable
(edge measurement and glitch measurements).