Buttons Updated--not stable
Buttons in good order. Need to add a temp pattern so that the new
pattern isn't always saved.

ERROR: the sinlon pattern changes brightness (with the adjustable
brightness turned off) and seems to roll over for brightness through out
the sweep.  It's unclear when this error was introduced but seems to
have been around for a while but hidden by the adjustable brightness
feature which was overriding the brightness change.